removed default counter store and created detections store.
