+++ dogfinder-app/src/stores/detections.ts
@@ -1,0 +1,12 @@
+import { ref, computed } from 'vue'
+import { defineStore } from 'pinia'
+
+export const useCounterStore = defineStore('counter', () => {
+        const count = ref(0)
+        const doubleCount = computed(() => count.value * 2)
+        function increment() {
+                count.value++
+        }
+
+        return { count, doubleCount, increment }
+})
